feat(projection): add projection.Checkpoint and projection.DoNotCheckpoint
